requirements.txt
pyyaml>=6
python>=3